+++ lib/tilt/string.rb
@@ -7,6 +7,7 @@ module Tilt
   class StringTemplate < Template
     def prepare
       hash = "TILT#{@data.hash.abs}"
+      @freeze_string_literals = !!@options[:freeze]
       @code = String.new("<<#{hash}.chomp\n#{@data}\n#{hash}")
     end
 
@@ -17,6 +18,10 @@ module Tilt
     def precompiled(locals)
       source, offset = super
       [source, offset + 1]
+    end
+
+    def freeze_string_literals?
+      @freeze_string_literals
     end
   end
 end