+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Remove deprecated support for RedCarpet 1.x (jeremyevans)
 * Remove deprecated support for Tilt.current_template (jeremyevans)
 * Make Template#freeze_string_literals? work correctly with Template#compiled_path (jeremyevans)
+* Support :freeze option for StringTemplate to support frozen string literals (jeremyevans)
 
 ## 2.2.0 (2023-06-05)